// frontend/src/app/core/models/candidate.model.ts
export interface Candidate {
  id?: number;
  firstName: string;
  lastName: string;
  email: string;
  phone?: string;
  username?: string;
  title?: string;
  skills?: string;
  experienceYears?: number;
  resumeUrl?: string;
  coverLetter?: string;
  address?: string;
  dateOfBirth?: string;
  linkedin?: string;
}